Move `PlayerInfoFlags`
