+++ pumpkin-protocol/Cargo.toml
@@ -32,6 +32,7 @@ cfb8 = "0.8"
 async-compression = { version = "0.4.20", features = ["tokio", "zlib"] }
 
 take_mut = "0.2.2"
+bitflags = "2.9.0"
 
 [dev-dependencies]
 # Validate correctness
